+++ src/pymdma/cli.py
@@ -16,6 +16,7 @@ from pymdma.constants import (
     DataModalities,
     EvaluationLevel,
     InputMetricGroups,
+    MetricGoal,
     ReferenceType,
     SyntheticMetricGroups,
     ValidationTypes,
@@ -57,7 +58,7 @@ def parse_args():
         type=str,
         nargs="+",
         default=None,
-        help="Metrics to be evaluated. E.g. feature, quality etc.",
+        help="Metrics to be evaluated. E.g. privacy, quality etc.",
     )
     parser.add_argument(
         "--metric_goals",
@@ -123,6 +124,12 @@ def parse_args():
         default=1,
         help="Number of workers to be used in the computation. Defaults to 1.",
     )
+    parser.add_argument(
+        "--device",
+        type=str,
+        default="cpu",
+        help="Device to be used for computation. Defaults to 'cpu'.",
+    )
     return parser.parse_args()
 
 
@@ -138,7 +145,7 @@ def infer_data_source(data_modality: str, data_path: Path):
         return data_path
 
     # modality custom data parsers
-    module = import_module(f"{data_modality}.data.parsers")
+    module = import_module(f"pymdma.{data_modality}.data.parsers")
     if data_path.suffix == ".jsonl":
         return module.jsonl_files(data_path)
 
@@ -183,6 +190,9 @@ def main() -> None:
             metric_goals=None,
         )
 
+    if args.annotation_file is None:
+        s_func.pop("annotation", None)
+
     for eval_group in list(s_func.keys()):
         funcs = s_func[eval_group]
         if len(funcs) == 0:
@@ -208,6 +218,7 @@ def main() -> None:
         args.batch_size,
         args.output_dir if args.allow_feature_cache else None,
         annotation_file=args.annotation_file,
+        device=args.device,
     )
 
     logger.info(
@@ -253,7 +264,11 @@ def main() -> None:
     with open(args.output_dir / "output.json", "w") as f:
         f.write(json.dumps(output, indent=2))
 
-    logger.info(f"Results saved to {args.output_dir / 'output.json'}")
+    with open(args.output_dir / "config.json", "w") as f:
+        args_vals = {key: str(val) for key, val in dict(vars(args)).items()}
+        json.dump(args_vals, f, indent=2)
+
+    logger.info(f"Results saved to {args.output_dir}")
 
 
 if __name__ == "__main__":