+++ src/pymdma/image/measures/input_val/annotation/coco.py
@@ -5,9 +5,7 @@ from pycocotools.coco import COCO
 
 from pymdma.common.definitions import Metric
 from pymdma.common.output import MetricResult
-from pymdma.constants import AnnotationType, EvaluationLevel
-from pymdma.constants import InputAnnotationMetrics as AnnotationMetrics
-from pymdma.constants import OutputsTypes, ReferenceType
+from pymdma.constants import AnnotationType, EvaluationLevel, MetricGoal, OutputsTypes, ReferenceType
 
 _SUPPORTED_ANNOT_TYPES = {
     "segmentation": AnnotationType.MASK,
@@ -62,7 +60,7 @@ class DatasetCompletness(Metric):
 
     reference_type = ReferenceType.NONE
     evaluation_level = [EvaluationLevel.DATASET, EvaluationLevel.INSTANCE]
-    metric_goal = AnnotationMetrics.COMPLETENESS
+    metric_goal = MetricGoal.VALIDITY
     annotation_type = [AnnotationType.LABEL, AnnotationType.BBOX, AnnotationType.MASK, AnnotationType.KEYPOINTS]
 
     def __init__(
@@ -189,7 +187,7 @@ class AnnotationCorrectness(Metric):
 
     reference_type = ReferenceType.NONE
     evaluation_level = [EvaluationLevel.DATASET, EvaluationLevel.INSTANCE]
-    metric_goal = AnnotationMetrics.CORRECTNESS
+    metric_goal = MetricGoal.VALIDITY
     annotation_type = [AnnotationType.LABEL, AnnotationType.BBOX, AnnotationType.MASK, AnnotationType.KEYPOINTS]
 
     def __init__(
@@ -373,7 +371,7 @@ class AnnotationUniqueness(Metric):
 
     reference_type = ReferenceType.NONE
     evaluation_level = [EvaluationLevel.DATASET, EvaluationLevel.INSTANCE]
-    metric_goal = AnnotationMetrics.UNIQUENESS
+    metric_goal = MetricGoal.VALIDITY
     annotation_type = [AnnotationType.BBOX, AnnotationType.MASK, AnnotationType.KEYPOINTS]
 
     def __init__(