+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.1.2] - 2024-10-28
+### Added
+ - CLI device specification
+ - CLI compute arguments in reports dir
+ - Preprocessing transform in image datasets when using the API
+
+### Fixed
+ - Moved `sklearn` and `gudhi` dependencies to the main dependency tree
+
+### Changed
+ - Default image feature extractor is now `vit_b_32`
+ - Confusing synthetic and input metric goals where aggregated to `quality`, `privacy`, `annotation` and `utility` categories
+ - Moved metrics to sepecific folders based on `metric_group` (feature-based, data-based)
+
+
 ## [0.1.1] - 2024-10-24
 ### Fixed
  - Fixed project configuration conflict between setup.py and pyproject.toml by reverting to poetry as main build engine